improve turbo mode support
- delay the loading of the module that we only need in the background
  process.
- ensure we've on single AnsibleModule instance
- flatten OKDRawModule class
- main() reuse the kubernetes.core logic
